feat: Tabbed Settings Dialog (General / Appearance / Shortcuts)
