+++ app/frontend/src/components/dialog.tsx
@@ -7,8 +7,11 @@ type DialogProps = {
   children: React.ReactNode;
   /** Width variant (260724-6j1v): `sm` (default) keeps the phone-card
    *  `max-w-sm` every existing dialog uses; `lg` is the desktop preference-pane
-   *  width (`max-w-2xl`, ≈672px) the settings dialog opts into. */
-  size?: "sm" | "lg";
+   *  width (`max-w-2xl`, ≈672px). `xl` (`max-w-4xl`) additionally pins a FIXED
+   *  panel height with `overflow-hidden` — the tabbed settings dialog's
+   *  variant, so a tab rail never jumps between short and tall panels (each
+   *  tab panel owns its own internal scroll instead). */
+  size?: "sm" | "lg" | "xl";
 };
 
 export function Dialog({ title, onClose, children, size = "sm" }: DialogProps) {
@@ -18,6 +21,15 @@ export function Dialog({ title, onClose, children, size = "sm" }: DialogProps) {
   // Dialog only mounts while open, so the trap is unconditionally active.
   // The hook owns focus-first-on-mount, Escape → onClose, and Tab wrap.
   useFocusTrap(dialogRef, true, onClose);
+
+  // max-h + overflow-y keep tall dialogs (the lg settings pane) scrollable
+  // inside short viewports instead of clipping off-screen; the calc offset
+  // matches the container's p-4 so the panel never touches the edges. xl
+  // inverts that: a fixed height with the scroll pushed into the tab panels.
+  const sizing =
+    size === "xl"
+      ? "max-w-4xl h-[min(40rem,calc(100vh-2rem))] overflow-hidden flex flex-col"
+      : `${size === "lg" ? "max-w-2xl" : "max-w-sm"} max-h-[calc(100vh-2rem)] overflow-y-auto`;
 
   return (
     <div
@@ -30,10 +42,7 @@ export function Dialog({ title, onClose, children, size = "sm" }: DialogProps) {
         role="dialog"
         aria-modal="true"
         aria-labelledby={titleId}
-        // max-h + overflow-y keep tall dialogs (the lg settings pane) scrollable
-        // inside short viewports instead of clipping off-screen; the calc offset
-        // matches the container's p-4 so the panel never touches the edges.
-        className={`relative bg-bg-primary border border-border rounded-lg p-3 w-full ${size === "lg" ? "max-w-2xl" : "max-w-sm"} max-h-[calc(100vh-2rem)] overflow-y-auto shadow-2xl text-[11px]`}
+        className={`relative bg-bg-primary border border-border rounded-lg p-3 w-full ${sizing} shadow-2xl text-[11px]`}
         onClick={(e) => e.stopPropagation()}
       >
         <h2 id={titleId} className="text-xs font-medium mb-2.5">{title}</h2>
