+++ app/frontend/tests/e2e/settings-dialog.spec.ts
@@ -6,14 +6,19 @@ import { gotoServerReady } from "./_ready";
 import { TMUX_SERVER, createSession, killSession, listWindows } from "./_tmux";
 
 /**
- * Settings dialog (260723-o7q8): the VS Code-style dialog mounted once at
- * AppLayout. These tests prove the four intake-level behaviors:
- *   1. palette-open on a server route with the This-host/This-device split,
+ * Settings dialog (260723-o7q8; TABBED in 260818-bncw — General / Appearance
+ * / Shortcuts, one tablist: left rail ≥480px, horizontal strip below): the
+ * VS Code-style dialog mounted once at AppLayout. These tests prove the
+ * intake-level behaviors:
+ *   1. palette-open on a server route — General by default (name/SSH/
+ *      notifications), Appearance carrying theme/accent/font,
  *   2. palette-open on /board/$name (the AppLayout mount's whole point —
  *      the board route renders no AppShell),
  *   3. top-bar gear open (the gear relocated from the sidebar footer to the
  *      top-bar right cluster in 260812-d1at),
  *   4. a host-scoped edit (instance name) persists through the API.
+ *   5. 375px: the horizontal tab strip fits, and the tall Shortcuts panel
+ *      scrolls INSIDE the fixed-height xl dialog.
  *
  * scripts/test-e2e.sh isolates the tmux server/port but NOT $HOME, so the
  * instance-name write lands in the developer's REAL ~/.rk/settings.yaml —
@@ -82,20 +87,29 @@ test.describe("Settings dialog", () => {
     killSession(TEST_SESSION);
   });
 
-  test("palette opens the dialog on a server route with the This-host/This-device split", async ({ page }) => {
+  test("palette opens the dialog on the General tab; the Appearance tab carries the rest (260818-bncw)", async ({ page }) => {
     await gotoServerReady(page, TMUX_SERVER);
 
     await openPaletteSettings(page);
     await expectDialogOpen(page);
 
-    // Scope split — both labeled sections render with their controls. Scope
-    // to the dialog: the sidebar HOST panel carries its own accent-picker
-    // button with the same accessible name.
+    // General (the tab-less default): instance name + SSH host under This
+    // host, notifications under This device. Scope to the dialog: the
+    // sidebar HOST panel carries its own accent-picker button with the same
+    // accessible name.
     const dialog = page.getByRole("dialog", { name: "Settings" });
+    await expect(dialog.getByRole("tab", { name: "General" })).toHaveAttribute(
+      "aria-selected",
+      "true",
+    );
     await expect(dialog.getByText("This host")).toBeVisible();
     await expect(dialog.getByText("This device")).toBeVisible();
     await expect(dialog.getByLabel("Instance name")).toBeVisible();
     await expect(dialog.getByLabel("SSH host")).toBeVisible();
+    await expect(dialog.getByText("Notifications", { exact: true })).toBeVisible();
+
+    // The Appearance tab keeps its own scope split.
+    await dialog.getByRole("tab", { name: "Appearance" }).click();
     await expect(dialog.getByRole("button", { name: "Set instance color" })).toBeVisible();
     await expect(dialog.getByLabel("Dark theme")).toBeVisible();
     await expect(dialog.getByRole("button", { name: "Increase terminal font" })).toBeVisible();
@@ -105,16 +119,27 @@ test.describe("Settings dialog", () => {
     await expect(page.getByRole("dialog", { name: "Settings" })).not.toBeVisible();
   });
 
-  test("desktop preference-pane layout with the Notifications row (260724-6j1v)", async ({ page }) => {
+  test("tabbed preference-pane layout with the Notifications row (260724-6j1v, 260818-bncw)", async ({ page }) => {
     await gotoServerReady(page, TMUX_SERVER);
 
     await openPaletteSettings(page);
     await expectDialogOpen(page);
     const dialog = page.getByRole("dialog", { name: "Settings" });
 
-    // Wide lg Dialog variant (~672px) instead of the phone-card max-w-sm.
-    await expect(dialog).toHaveClass(/max-w-2xl/);
+    // The xl Dialog variant (max-w-4xl + fixed height) — wider than the old
+    // lg pane so the left rail + panel fit side by side.
+    await expect(dialog).toHaveClass(/max-w-4xl/);
     await expect(dialog).not.toHaveClass(/max-w-sm/);
+
+    // One tablist markup — the desktop left rail (a vertical tablist at this
+    // viewport) with roving tabindex (the active tab is the one Tab stop).
+    const tablist = dialog.getByRole("tablist", { name: "Settings sections" });
+    await expect(tablist).toBeVisible();
+    await expect(tablist.getByRole("tab", { name: "General" })).toHaveAttribute("tabindex", "0");
+    await expect(tablist.getByRole("tab", { name: "Appearance" })).toHaveAttribute(
+      "tabindex",
+      "-1",
+    );
 
     // Preference-row grid: each setting is a `190px 1fr` two-column grid at
     // desktop width (label column left, control column right — one vertical
@@ -125,9 +150,9 @@ test.describe("Settings dialog", () => {
     expect(rowClass).toContain("min-[480px]:grid-cols-[190px_1fr]");
 
     // Notifications row (moved from the retired top-bar bell) lives under the
-    // This-device scope: label, subscribed-gated test button, and the setup
-    // guide link. Status text varies by browser permission state, so only the
-    // state-independent contents are asserted.
+    // This-device scope on the General tab: label, subscribed-gated test
+    // button, and the setup guide link. Status text varies by browser
+    // permission state, so only the state-independent contents are asserted.
     await expect(dialog.getByText("Notifications", { exact: true })).toBeVisible();
     await expect(dialog.getByRole("button", { name: "Send test notification" })).toBeVisible();
     const guide = dialog.getByRole("link", { name: /Setup & troubleshooting guide/ });
@@ -139,10 +164,10 @@ test.describe("Settings dialog", () => {
     await expect(dialog).not.toBeVisible();
   });
 
-  test("short viewport (375x667): the dialog fits and its last row is reachable by scroll (260724-6j1v)", async ({ page }) => {
-    // The lg settings pane grew taller than a phone-landscape/short viewport;
-    // the Dialog panel must cap its height and scroll instead of clipping
-    // off-screen with no scroll path (rework finding M1).
+  test("short viewport (375x667): the tab strip fits and the tall Shortcuts panel scrolls internally (260724-6j1v, 260818-bncw)", async ({ page }) => {
+    // The xl dialog pins a fixed height with panel-level `overflow-hidden` —
+    // each tab panel owns its own scroll, so a tall tab must scroll inside
+    // the dialog instead of clipping off-screen with no scroll path.
     await page.setViewportSize({ width: 375, height: 667 });
     await page.goto(`/${TMUX_SERVER}`);
     // No [aria-label='Connected'] gate here: the dot lives in the sidebar
@@ -164,19 +189,32 @@ test.describe("Settings dialog", () => {
     expect(box!.x + box!.width).toBeLessThanOrEqual(375);
     expect(box!.y + box!.height).toBeLessThanOrEqual(667);
 
-    // The panel itself is the scroll container (its content overflows) …
-    const overflows = await dialog.evaluate((el) => el.scrollHeight > el.clientHeight);
-    expect(overflows).toBe(true);
+    // The same tablist markup is the horizontal strip under the title at
+    // <480px — all three tabs reachable.
+    for (const label of ["General", "Appearance", "Shortcuts"]) {
+      await expect(dialog.getByRole("tab", { name: label })).toBeVisible();
+    }
+    // … and the page itself never gains horizontal overflow.
+    const pageOverflow = await page.evaluate(
+      () => document.documentElement.scrollWidth > document.documentElement.clientWidth,
+    );
+    expect(pageOverflow).toBe(false);
 
-    // … and the LAST row's control (the Notifications setup-guide link) is
-    // reachable by scrolling within the panel.
-    const guide = dialog.getByRole("link", { name: /Setup & troubleshooting guide/ });
-    await guide.scrollIntoViewIfNeeded();
-    await expect(guide).toBeVisible();
-    const guideBox = await guide.boundingBox();
-    expect(guideBox).toBeTruthy();
-    expect(guideBox!.y).toBeGreaterThanOrEqual(0);
-    expect(guideBox!.y + guideBox!.height).toBeLessThanOrEqual(667);
+    // The Shortcuts tab is the tall panel: its TAB PANEL (not the dialog)
+    // overflows and scrolls, and its last element (the reset-all footer) is
+    // reachable by scrolling within it.
+    await dialog.getByRole("tab", { name: "Shortcuts" }).click();
+    await expect(page.getByTestId("settings-shortcuts-panel")).toBeVisible();
+    const tabpanel = dialog.getByRole("tabpanel");
+    const overflows = await tabpanel.evaluate((el) => el.scrollHeight > el.clientHeight);
+    expect(overflows).toBe(true);
+    const resetAll = dialog.getByRole("button", { name: "reset all" });
+    await resetAll.scrollIntoViewIfNeeded();
+    await expect(resetAll).toBeVisible();
+    const resetBox = await resetAll.boundingBox();
+    expect(resetBox).toBeTruthy();
+    expect(resetBox!.y).toBeGreaterThanOrEqual(0);
+    expect(resetBox!.y + resetBox!.height).toBeLessThanOrEqual(667);
   });
 
   test("palette opens the same dialog on /board/$name (no AppShell there)", async ({ page }) => {
@@ -199,6 +237,24 @@ test.describe("Settings dialog", () => {
       const dialog = page.getByRole("dialog", { name: "Settings" });
       await expect(dialog.getByText("This host")).toBeVisible();
       await expect(dialog.getByText("This device")).toBeVisible();
+
+      // The board shell's `shortcuts-overlay` chord handler resolves the same
+      // global entry: ⇧Ctrl+/ switches the open dialog to the Shortcuts tab
+      // here too, and re-firing closes (260818-bncw three-state toggle).
+      await page.keyboard.press("Shift+Control+Slash");
+      await expect(dialog.getByRole("tab", { name: "Shortcuts" })).toHaveAttribute(
+        "aria-selected",
+        "true",
+      );
+      await expect(page.getByTestId("settings-shortcuts-panel")).toBeVisible();
+      // Board route: no session context — the add flow stays gated off and
+      // the TMUX section renders its empty state, not a crash/spinner.
+      await expect(
+        page.getByText("+ bind a key to a palette action or riff preset…"),
+      ).toHaveCount(0);
+      await expect(page.getByTestId("tmux-section").getByText("No tmux server running")).toBeVisible();
+      await page.keyboard.press("Shift+Control+Slash");
+      await expect(dialog).toHaveCount(0);
     } finally {
       // Unpin so the board (and its _rk-pin-* session) does not outlive the run.
       await page.request
