+++ app/frontend/tests/e2e/top-bar-overflow.spec.md
@@ -184,15 +184,17 @@ in-bar entry's rows are NOT duplicated into the menu.
 ### `the App-section chrome rows work: Help links out, Keyboard opens the overlay, Theme… opens the selector`
 
 **What it proves:** the relocated rows (260812-d1at) are functional, not just
-present — Help is a safe external link, Keyboard shortcuts opens the
-ShortcutsOverlay, and Theme… opens the theme selector (the retired footer
+present — Help is a safe external link, Keyboard shortcuts deep-links into
+the settings dialog's Shortcuts tab (260818-bncw — the standalone overlay and
+its `shortcuts-overlay:open` event seam are retired), and Theme… opens the
+theme selector (the retired footer
 button's click-cycling is gone).
 
 **Steps:**
 1. Navigate to the terminal window; set 375×800; open the `More controls` menu.
 2. Assert the Help row's `href` / `target="_blank"` / `rel="noopener…"`.
-3. Click `Keyboard shortcuts`; assert the `shortcuts-overlay` testid is
-   visible; Escape-close it.
+3. Click `Keyboard shortcuts`; assert the `Settings` dialog is visible with
+   the `settings-shortcuts-panel` testid inside; Escape-close it.
 4. Reopen the menu; click `Theme…`; assert the `Theme selector` dialog is
    visible; Escape-close it.
