+++ docs/memory/run-kit/desktop-shell.md
@@ -298,7 +298,7 @@ The unshifted Ctrl tier is **entirely unbound** — the page tier there is compl
 
 The Hosts switcher **does** diverge, and by design: the radios bind `Shift+Ctrl+1–9` here — the win/linux shell tier — against `Alt+Cmd+1–9` on mac, capped by `MAX_SWITCHER_ACCELERATORS` on both. The keys pressed here are exactly what the former single `Shift+CmdOrCtrl` expression resolved to, so win/linux behavior is byte-identical and the split is representation-only on this side.
 
-**Two renderer-side copies mirror these bindings.** `menu.ts` is the *binding* authority, but the same mac-vs-other knowledge is restated for display in the SPA — `shortcuts-overlay.tsx` (`switcherCaps`) and `lib/shell-strip.ts` (`hostAcceleratorHint`, the strip dropdown's row hints, capped at 9 to match) — in deliberately different shapes across the main/renderer boundary. A change to the binding tier or the cap must update all three ([ui/top-bar](/run-kit/ui/top-bar.md) § Design Decisions → Host-switcher accelerator knowledge lives in three shapes).
+**Two renderer-side copies mirror these bindings.** `menu.ts` is the *binding* authority, but the same mac-vs-other knowledge is restated for display in the SPA — `settings-shortcuts-panel.tsx` (`switcherCaps`) and `lib/shell-strip.ts` (`hostAcceleratorHint`, the strip dropdown's row hints, capped at 9 to match) — in deliberately different shapes across the main/renderer boundary. A change to the binding tier or the cap must update all three ([ui/top-bar](/run-kit/ui/top-bar.md) § Design Decisions → Host-switcher accelerator knowledge lives in three shapes).
 
 Accepted gap (recorded nice-to-have): switcher radios check by exact `activeId` match, so in the dangling-`activeId` state startup loads the first host (store fallback) while no radio renders checked until the next list mutation rebuilds the menu.
