+++ docs/memory/run-kit/ui/index.md
@@ -9,9 +9,9 @@ description: "Frontend UI patterns, split by surface: routes + shell, top bar, s
 |------|-------------|
 | [boards](boards.md) | Boards: link-based pinning, board page layout, pane cards, kill confirm, autofit, pane reorder, relay-connection suspension, mobile carousel, sidebar boards section, pin entry points, board/pin/view palette actions, frontend hooks. |
 | [compose-and-bottom-bar](compose-and-bottom-bar.md) | Bottom bar (coarse-pointer-only) and docked compose strip: pointer-gated bar existence, send semantics and the Enter classifier, drafts + sent history, readline editing chords, file upload, iOS keyboard support, the card/compact compose layout with per-pointer morph triggers, edge-docked padding, the focus-memory recording seam. |
-| [dialogs-and-state](dialogs-and-state.md) | Component conventions, dialogs (create session, spawn-agent, settings, width variants), clipboard utility, e2e host-global filesystem state, the Zustand window store, and optimistic UI + mutation feedback. |
+| [dialogs-and-state](dialogs-and-state.md) | Component conventions, dialogs (create session, spawn-agent, tabbed settings, width variants), clipboard utility, e2e host-global filesystem state, the Zustand window store, and optimistic UI + mutation feedback. |
 | [focus-ownership](focus-ownership.md) | Per-window focus memory (tty/compose/code) and the code-server steal guard: real-focus-only recording asymmetry, the desktop restore router with its tty first-visit default, the guard arm/disarm lifecycle, the in-frame focusin grab detector, and the focus-hop chord (⌃`) as a focus-seam consumer that records only on the tty hop. |
-| [keyboard-and-palette](keyboard-and-palette.md) | Keyboard shortcuts: tier system and the e.code rule, default bindings, macros, override layer, claimed keys, shortcuts overlay, education micro-copy; command palette single mount, action registry, mobile trigger, and the tmux commands dialog. |
+| [keyboard-and-palette](keyboard-and-palette.md) | Keyboard shortcuts: tier system and the e.code rule, default bindings, macros, override layer, claimed keys, the shortcuts panel (the settings dialog's Shortcuts tab), education micro-copy; command palette single mount, action registry, mobile trigger, and the tmux commands dialog. |
 | [lenses-and-layout](lenses-and-layout.md) | The window-view lens model (tty/web/chat/desktop), the surface layout manager (shapes × surfaces × ratios, ?layout= ladder, gap-seam tile chrome), the rk present auto-expand transient reaction (renderLayout override + dismissal latch), top-bar surface toggles (desktop toggle + pinned mobile switch modes) and the switch-to-tile verb, iframe/code/chat surface renderers, the code-folder latch, and the tile renderer's focus-memory recording seams + steal-guard consultation. |
 | [routes-and-shell](routes-and-shell.md) | Route surface and app shell: URL structure as resumable bookmark, server-gone flip, session tiles, shell grid layout, live safe-name conversion, session creation flows, instance display name, session-to-project mapping, activity status. |
 | [sidebar](sidebar.md) | Sidebar: row anatomy, row identity hover tips, icon system and label zone, the three-tier coarse status rail + cards, row flair overlays, render-performance constraints, keyboard nav + tree ARIA, multi-select + broadcast, collapsible panels, section-visibility rail, host-page zones (health/boards/services), kill controls, scroll-edge fade, operator pinned row, footer. |
