+++ fab/changes/260818-bncw-tabbed-settings-dialog/.history.jsonl
@@ -1,0 +1,10 @@
+{"action":"enter","driver":"fab-new","event":"stage-transition","stage":"intake","ts":"2026-08-18T00:44:48Z"}
+{"args":"Consolidate all settings surfaces into a single tabbed settings dialog (General/Appearance/Shortcuts) with left nav rail on desktop, horizontal tab strip under 480px; fold the shortcuts overlay body in as the Shortcuts tab; openSettings(tab?) deep-links from both chords and per-tab palette actions","cmd":"fab-new","event":"command","ts":"2026-08-18T00:44:48Z"}
+{"delta":"+4.3","event":"confidence","score":4.3,"trigger":"calc-score","ts":"2026-08-18T00:48:13Z"}
+{"cmd":"fab-switch","event":"command","ts":"2026-08-18T00:49:30Z"}
+{"cmd":"fab-fff","event":"command","ts":"2026-08-18T00:50:53Z"}
+{"action":"enter","driver":"fab-fff","event":"stage-transition","stage":"apply","ts":"2026-08-18T00:54:06Z"}
+{"action":"enter","driver":"fab-fff","event":"stage-transition","stage":"review","ts":"2026-08-18T03:10:49Z"}
+{"action":"enter","driver":"fab-fff","event":"stage-transition","stage":"hydrate","ts":"2026-08-18T03:17:41Z"}
+{"event":"review","result":"passed","ts":"2026-08-18T03:17:41Z"}
+{"action":"enter","driver":"fab-fff","event":"stage-transition","stage":"ship","ts":"2026-08-18T03:33:04Z"}
