+++ fab/changes/260818-bncw-tabbed-settings-dialog/.status.yaml
@@ -1,0 +1,54 @@
+id: bncw
+name: 260818-bncw-tabbed-settings-dialog
+created: 2026-08-18T00:44:48Z
+created_by: sahil-noon
+change_type: feat
+issues: []
+progress:
+    intake: done
+    apply: done
+    review: done
+    hydrate: done
+    ship: active
+    review-pr: pending
+plan:
+    generated: true
+    task_count: 10
+    acceptance_count: 17
+    acceptance_completed: 17
+confidence:
+    certain: 9
+    confident: 6
+    tentative: 0
+    unresolved: 0
+    score: 4.3
+    fuzzy: true
+    dimensions:
+        signal: 75.0
+        reversibility: 78.7
+        competence: 82.3
+        disambiguation: 80.7
+stage_metrics:
+    intake: {started_at: "2026-08-18T00:44:48Z", driver: fab-new, iterations: 1, completed_at: "2026-08-18T00:54:06Z"}
+    apply: {started_at: "2026-08-18T00:54:06Z", driver: fab-fff, iterations: 1, completed_at: "2026-08-18T03:10:49Z"}
+    review: {started_at: "2026-08-18T03:10:49Z", driver: fab-fff, iterations: 1, completed_at: "2026-08-18T03:17:41Z"}
+    hydrate: {started_at: "2026-08-18T03:17:41Z", driver: fab-fff, iterations: 1, completed_at: "2026-08-18T03:33:04Z"}
+    ship: {started_at: "2026-08-18T03:33:04Z", driver: fab-fff, iterations: 1}
+prs: []
+change_type_source: explicit
+true_impact:
+    added: 0
+    deleted: 0
+    net: 0
+    excluding:
+        added: 0
+        deleted: 0
+        net: 0
+    tests:
+        added: 0
+        deleted: 0
+        net: 0
+    computed_at: "2026-08-18T03:33:04Z"
+    computed_at_stage: hydrate
+# true_impact: lazily created on first stage-finish that computes it (no placeholder here).
+last_updated: 2026-08-18T03:33:04Z
